feat: add bunked status and improve system stability
- Add BUNKED status to AttendanceStatus and update calculators
- Implement DataStore error handling in UserPreferencesRepository
- Add lifecycle management for BroadcastReceiver tasks
- Add timetable export utilities

+++ app/src/main/java/com/agupta07505/attendmate/ui/screens/subjectdetails/SubjectDetailScreen.kt
@@ -57,7 +57,7 @@ fun SubjectDetailScreen(
             }
         } else {
             val sub = subject!!
-            val subjectColor = Color(sub.colorValue.toULong())
+            val subjectColor = Color(sub.colorValue.toInt())
 
             LazyColumn(
                 modifier = Modifier
